# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.AuthResponse:
    properties:
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/schemas.UserResponse'
    type: object
...
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: joao@example.com
        type: string
      password:
        example: senha123
        type: string
    required:
    - email
    - password
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        example: joao@example.com
        type: string
      name:
        example: João Silva
        minLength: 2
        type: string
      password:
        example: senha123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  schemas.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
host: 147.185.221.212:61489
info:
  contact:
    email: suporte@exemplo.com
    name: Equipe de Desenvolvimento
    url: https://github.com/Pmmvito/Golang-Api-Exemple
  description: "API REST para autenticação JWT, análise inteligente de notas fiscais
    de supermercado e gestão de categorias\n\n## \U0001F680 Funcionalidades Principais:\n-
    \U0001F510 **Autenticação JWT** com tokens de 7 dias\n- \U0001F916 **Análise de
    Notas com IA** usando Google Gemini (2-3s)\n- \U0001F4CA **Categorização Automática**
    de produtos de supermercado\n- \U0001F4DD **CRUD Completo** de recibos e categorias\n-
    \U0001F50D **Filtros Avançados** por categoria, data e valor\n- ⚡ **Respostas
    Otimizadas** (55% menores)\n\n## \U0001F4D6 Como Usar:\n1. Registre-se em `/register`\n2.
    Faça login em `/login` para obter o token JWT\n3. Clique em **Authorize** \U0001F513
    (cadeado verde) e cole: `Bearer SEU_TOKEN`\n4. Agora você pode testar todos os
    endpoints protegidos!\n\n## \U0001F6D2 Categorias de Supermercado (21 categorias
    distintas):\n**Básicos:** Grãos e Cereais, Massas, Padaria\n**Proteínas:** Carnes
    e Proteínas, Frios e Embutidos\n**Laticínios:** Leite, Queijos, Iogurtes, Manteiga\n**Frescos:**
    Frutas e Vegetais\n**Bebidas:** Bebidas (não alcoólicas), Bebidas Alcoólicas,
    Café e Chá\n**Congelados:** Produtos Congelados\n**Doces:** Doces e Sobremesas,
...
    Higiene Pessoal, Limpeza Doméstica, Papel e Descartáveis\n**Especiais:** Bebê
...
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password. Returns a JWT token
        valid for 7 days. Use this token in the Authorization header as "Bearer {token}"
        for all protected endpoints.
      parameters:
      - description: User credentials (email and password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - "\U0001F510 Authentication"
...
  /me:
    get:
      consumes:
      - application/json
      description: Get information about the authenticated user (requires JWT token)
      produces:
      - application/json
      responses:
        "200":
          description: User information retrieved successfully
          schema:
            $ref: '#/definitions/schemas.UserResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - "\U0001F510 Authentication"
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account. After registration, use the login endpoint
        to get your JWT token.
      parameters:
      - description: User registration data (name, email, password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully with JWT token
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Invalid request or email already registered
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register new user
      tags:
      - "\U0001F510 Authentication"
...
  /user:
    delete:
      consumes:
      - application/json
      description: Soft delete user account and all associated data (receipts, items,
        products, shopping lists, tokens). Email cannot be reused even after deletion.
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer" seguido do seu token JWT (obtido no login). Exemplo:
      Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"